+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased][]
 
+- Added: Dialogs
+- Fixed: HandleScrollInput allowing scrolling behind
+- Fixed: ApplyAncestorClips' content overflow
+
 ## v[3.0.0][] 2026-09-18
 
 - Breaking CHANGE: integrations namespaces now use only `Guinevere`